+++ Tunny/Settings/Random.cs
@@ -1,0 +1,10 @@
+namespace Tunny.Settings
+{
+    /// <summary>
+    /// https://optuna.readthedocs.io/en/stable/reference/generated/optuna.samplers.RandomSampler.html#optuna.samplers.RandomSampler
+    /// </summary>
+    public class Random
+    {
+        public int? Seed { get; set; }
+    }
+}